COM-1316 extract slots and trainees management out of profileOrganization
